Adapt benchmarks function to new `VM`'s interface
- Change `CARGO_TARGET_DIR`.
- Move helper functions from `tests.rs` to new `utils.rs`.
