+++ crates/levm/bench/revm_comparison/README.md
@@ -1,7 +1,7 @@
 # Benchmarks
 This README explains how to run benchmarks to compare the performance of `levm` and `revm` when running different contracts. The benchmarking tool used to gather performance metrics is [hyperfine](https://github.com/sharkdp/hyperfine), and the obtained results will be included for reference.
 
-To run the benchmarks (from the crate's root):
+To run the benchmarks (from `levm`'s root):
 ```bash
 make revm-comparison
 ```
